se inicia uso de ajax, revisar video 34

+++ resources/views/tickets/details.blade.php
@@ -30,21 +30,24 @@
                 <span class="label label-info">{{ $user->name }}</span>
                 @endforeach
             </p>
-
-            @if( ! currentUser()->hasVoted($ticket))
-            {!! Form::open(['route' => ['votes.submit', $ticket->id], 'method' => 'POST']) !!}
+            @if(Auth::check())
+                @if( ! currentUser()->hasVoted($ticket))
+                {!! Form::open(['route' => ['votes.submit', $ticket->id], 'method' => 'POST']) !!}
+                    <button type="submit" class="btn btn-primary">
+                        <span class="glyphicon glyphicon-thumbs-up"></span> Votar
+                    </button>
+                {!! Form::close() !!}
+                @else
+                {!! Form::open(['route' => ['votes.destroy', $ticket->id], 'method' => 'DELETE']) !!}
                 <button type="submit" class="btn btn-primary">
-                    <span class="glyphicon glyphicon-thumbs-up"></span> Votar
+                    <span class="glyphicon glyphicon-thumbs-down"></span>Quitar Votar
                 </button>
-            {!! Form::close() !!}
-            @else
-            {!! Form::open(['route' => ['votes.destroy', $ticket->id], 'method' => 'DELETE']) !!}
-            <button type="submit" class="btn btn-primary">
-                <span class="glyphicon glyphicon-thumbs-down"></span>Quitar Votar
-            </button>
-            {!! Form::close() !!}
+                {!! Form::close() !!}
+                @endif
             @endif
-            <h3>Nuevo Comentario</h3>
+
+
+                <h3>Nuevo Comentario</h3>
 
             @include('partials/errors')
